+++ src/core/hle/kernel/kernel.cpp
@@ -156,7 +156,7 @@ struct KernelCore::Impl {
     // TODO(Subv): Start the process ids from 10 for now, as lower PIDs are
     // reserved for low-level services
     std::atomic<u64> next_process_id{10};
-    std::atomic<u32> next_thread_id{1};
+    std::atomic<u64> next_thread_id{1};
 
     // Lists all processes that exist in the current session.
     std::vector<SharedPtr<Process>> process_list;
@@ -242,7 +242,7 @@ u32 KernelCore::CreateNewObjectID() {
     return impl->next_object_id++;
 }
 
-u32 KernelCore::CreateNewThreadID() {
+u64 KernelCore::CreateNewThreadID() {
     return impl->next_thread_id++;
 }
